Basic save and remove

+++ github-projects.js
@@ -2,28 +2,64 @@ var $pagehead_actions = document.querySelector('.pagehead-actions');
 var key = 'github-projects';
 var repos = JSON.parse(localStorage.getItem(key)) || [];
 
+function already_saved(url) {
+  var in_array = false;
+
+  repos.forEach(function(repo) {
+    if (repo.url == url) {
+      in_array = true;
+    }
+  });
+
+  return in_array;
+}
+
 console.log(repos);
 
 if ($pagehead_actions) {
-  var $add_repo = document.createElement('button');
-  $add_repo.textContent = "Save Repo";
-  $add_repo.classList.add('btn');
-  $add_repo.classList.add('btn-sm');
+  var url = window.location.href;
+  var $button;
 
-  $add_repo.addEventListener('click', function() {
-    var repo = {
-      url: window.location.href
-    };
-    repos.push(repo);
+  if (already_saved(url)) {
+    $button = document.createElement('button');
+    $button.textContent = "Remove Repo";
+    $button.classList.add('btn');
+    $button.classList.add('btn-sm');
 
-    localStorage.setItem(key, JSON.stringify(repos));
+    $button.addEventListener('click', function() {
+      var index;
 
-    alert('repo saved!')
-  });
+      repos.forEach(function(repo, i) {
+        if (repo.url == url) {
+          index = i;
+        }
+      });
+
+      repos.splice(index, 1);
+      localStorage.setItem(key, JSON.stringify(repos));
+
+      alert('repo remove!');
+    });
+  } else {
+    $button = document.createElement('button');
+    $button.textContent = "Save Repo";
+    $button.classList.add('btn');
+    $button.classList.add('btn-sm');
+
+    $button.addEventListener('click', function() {
+      var repo = {
+        url: url
+      };
+      repos.push(repo);
+
+      localStorage.setItem(key, JSON.stringify(repos));
+
+      alert('repo saved!');
+    });
+  }
 
   var $li = document.createElement('li');
-  $li.appendChild($add_repo);
-
+  $li.appendChild($button);
   $pagehead_actions.appendChild($li);
 }
